refactor: 공통 dto 폴더 구조 개선
